+++ src/postProcessingStage.c
@@ -25,7 +25,7 @@ SOFTWARE.
 #include "postProcessingStage.h"
 static ring_buffer_t *peakBuf;
 static data_point_t lastDataPoint;
-static int timeThreshold = 200;
+static int16_t timeThreshold = 200;
 static void (*stepCallback)(void);
 void initPostProcessingStage(ring_buffer_t *peakBufIn, void (*stepCallbackIn)(void))
 {
@@ -69,7 +69,7 @@ void resetPostProcess(void)
     lastDataPoint.time = 0;
 }
 
-void changeTimeThreshold(int thresh)
+void changeTimeThreshold(int16_t thresh)
 {
     timeThreshold = thresh;
 }
