+++ src/preProcessingStage.c
@@ -26,10 +26,10 @@ SOFTWARE.
 #include "utils.h"
 static ring_buffer_t *rawBuf;
 static ring_buffer_t *ppBuf;
-static int interpolationTime = 10;       //in ms
-static float timeScalingFactor = 1; //(100000 for validation data) 1000000
-static float startTime = -1;
-static int interpolationCount = 0;
+static int8_t interpolationTime = 10;       //in ms
+static int16_t timeScalingFactor = 1; //(100000 for validation data) 1000000
+static int32_t startTime = -1;
+static int16_t interpolationCount = 0;
 
 //static inline long labs(long num)
 //{
@@ -42,7 +42,7 @@ void initPreProcessStage(ring_buffer_t *rawBufIn, ring_buffer_t *ppBufIn)
     ppBuf = ppBufIn;
 }
 
-static data_point_t linearInterpolate(data_point_t dp1, data_point_t dp2, long interpTime)
+static data_point_t linearInterpolate(data_point_t dp1, data_point_t dp2, int64_t interpTime)
 {
     long mag = (dp1.magnitude + ((dp2.magnitude - dp1.magnitude) / (dp2.time - dp1.time)) * (interpTime - dp1.time));
     data_point_t interp;
@@ -51,14 +51,14 @@ static data_point_t linearInterpolate(data_point_t dp1, data_point_t dp2, long i
     return interp;
 }
 
-void preProcessSample(long time, long x, long y, long z)
+void preProcessSample(int64_t time, int32_t x, int32_t y, int32_t z)
 {
     if (startTime == -1) //first point handler
     {
         startTime = time;
     }
     //long magnitude = (labs(x * x) + labs(y * y) + labs(z * z)) / (labs(x) + labs(y) + labs(z));
-    long magnitude = isqrt(x*x+y*y+z*z); //Original
+    int64_t magnitude = isqrt((int64_t)(x*x+y*y+z*z)); //Original
     //long magnitude = (long)sqrt((x*x)+(y*y)+(z*z));
     data_point_t dataPoint;
     dataPoint.time = (time - startTime) / timeScalingFactor;
@@ -70,10 +70,10 @@ void preProcessSample(long time, long x, long y, long z)
         data_point_t dp2;
         ring_buffer_peek(rawBuf, &dp1, 0);
         ring_buffer_peek(rawBuf, &dp2, 1);
-        int numberOfPoints = 1 + ((((dp2.time - dp1.time)) - 1) / interpolationTime); //celing function
-        for (int i = 0; i < numberOfPoints; i++)
+        int16_t numberOfPoints = 1 + ((((dp2.time - dp1.time)) - 1) / interpolationTime); //celing function
+        for (int16_t i = 0; i < numberOfPoints; i++)
         {
-            long interpTime = (long)interpolationCount * interpolationTime;
+            int64_t interpTime = (int64_t)interpolationCount * interpolationTime;
             if (dp1.time <= interpTime && interpTime < dp2.time)
             {
                 data_point_t interpolated = linearInterpolate(dp1, dp2, interpTime);
